+++ RetroFE/Source/Graphics/PageBuilder.h
@@ -22,7 +22,7 @@
 #include <rapidxml.hpp>
 #include <vector>
 
-static const int MENU_INDEX_HIGH = 16;
+static const size_t MENU_INDEX_HIGH = 16;
 
 class ScrollingList;
 class Page;
